cid#318912 Result is not floating-point
we presumably want the current behaviour so silence with
an appropiate explicit cast

+++ wsd/ClientSession.cpp
@@ -517,7 +517,7 @@ bool ClientSession::_handleInput(const char *buffer, int length)
                 endptr = nullptr;
                 double counter = strtod(str, &endptr);
                 if (*endptr == '\0' && counter > 0 &&
-                    (counter < (double)(std::numeric_limits<uint64_t>::max() / 1000)))
+                    (counter < (uint64_t)(std::numeric_limits<uint64_t>::max() / 1000)))
                 {
                     // Now we know how to translate from the client's performance.now() values to
                     // microseconds since the epoch.
